requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.30.0
prisma>=0.15.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0
google-adk>=0.3.0
